requirements.txt
# Web/app
streamlit==1.35.0

# OpenAI
openai==1.26.0

# Email
yagmail==0.15.293

# Database
pymongo==4.13.1
dnspython==2.7.0

# Data/Date
python-dotenv==1.0.1
pandas==2.3.0
numpy==1.26.4
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# Visualization (if needed)
altair==5.5.0
pydeck==0.9.1
pillow==10.4.0  # Only if using images in Streamlit

# Utilities (used by dependencies)
attrs==25.3.0
click==8.2.1
requests==2.32.4
tenacity==8.5.0
typing_extensions==4.14.0
watchdog==6.0.0